✨ feat(student): add grade information to student model
The changes made in this commit add the `gradeResModel` property to the `StudentResModel` class. This property is used to store the grade information for the student.

The `toJson()` method of the `StudentResModel` class has also been updated to include the `grades` key in the JSON data, which contains the serialized `GradeResModel` object.

Additionally, the `fromJson()` method has been updated to deserialize the `grades` key from the JSON data and populate the `gradeResModel` property accordingly.

These changes will allow the application to better represent the relationship between students and their grades, making it easier to work with and display this information.