🎨 feat(flutter_launcher_icons): Add web icon configuration
Adds a new web icon configuration to the `flutter_launcher_icons.yaml`
file. The new configuration sets the `generate` flag to `true` and
specifies the image path for the web icon as `"assets/logos/NIS5.png"`.
This change allows for the web platform to have a custom icon
generated during the build process.