🎨 feat(routes): Add new route for students in exam room
Adds a new route constant `studentsInExamRoom` to the `Routes` class in `lib/routes_manager.dart`. This new route will be used to navigate to the screen that displays the list of students currently in the exam room.

The changes also include reordering the route constants for better organization and readability.