🩹 feat: Add and remove student cheating actions
This commit adds new constants for marking and unmarking a student as cheating, and removes
the unused constants for marking and unmarking a student as cheating. These changes are
made to the `StudentsLinks` class in the `app_links.dart` file.

The new constants are:
- `markCheatingStudent`
- `unMarkCheatingStudent`

These changes are made to ensure that the application's links are up-to-date and
accurately reflect the available functionality.