✨ feat: Update CanvasKit assets and last build ID
This commit updates the CanvasKit assets, including the `canvaskit.wasm` and
`canvaskit.js.symbols` files, as well as the last build ID for the
`control-proctor-app-2025` Firebase hosting environment.

These changes are necessary to ensure the latest version of the CanvasKit
library is being used, which provides important performance and functionality
improvements for the application.