🔖 feat(profile): Update cached user profile in Hive
This change updates the `saveProfileToHiveBox()` method in the `ProfileController`
to first assign the `cachedUserProfile` to the `_cachedUserProfile` property before
saving it to the Hive box. This ensures that the `update()` method is called after
the profile has been updated, triggering a rebuild of the UI widgets that depend
on this controller.