🔌 feat(version): Increase version to 1.1.11
This commit updates the version number in the `pubspec.yaml` file from 1.1.10 to 1.1.11. This change is necessary to reflect the latest version of the application.

Additionally, this commit adds a new route handling for the `unknown` route, which will redirect the user to the `LoginForm` screen. This change ensures that if the user navigates to an unknown route, they are seamlessly redirected to the login screen, providing a better user experience.